Separate out test classes into individual files
